fixup! simplified plugin
